+++ mountpoint-s3-client/tests/head_object.rs
@@ -47,6 +47,7 @@ async fn test_head_object() {
     );
 }
 
+#[test_case(ChecksumAlgorithm::Crc64Nvme)]
 #[test_case(ChecksumAlgorithm::Crc32)]
 #[test_case(ChecksumAlgorithm::Crc32C)]
 #[test_case(ChecksumAlgorithm::Sha1)]
@@ -83,6 +84,10 @@ async fn test_head_object_checksum(checksum_algorithm: ChecksumAlgorithm) {
         let checksum = result.checksum;
         if retrieve_checksum {
             match &checksum_algorithm {
+                ChecksumAlgorithm::Crc64Nvme => assert_eq!(
+                    checksum.checksum_crc64nvme,
+                    put_object_output.checksum_crc64_nvme().map(|s| s.to_string())
+                ),
                 ChecksumAlgorithm::Crc32 => assert_eq!(
                     checksum.checksum_crc32,
                     put_object_output.checksum_crc32().map(|s| s.to_string())