+++ mountpoint-s3-crt-sys/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+## v0.12.0 (February 3, 2025)
+
+* Update to latest CRT dependencies.
+
+## v0.11.0 (January 20, 2025)
+
 * Update to latest CRT dependencies
 
 ## v0.10.0 (October 17, 2024)